Task3_ImageCaptioning/requirements.txt
torch
torchvision
matplotlib
pillow
numpy
nltk
tqdm
